requirements.txt
flask>=3.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
tweepy>=4.14.0
openpyxl>=3.1.0
requests>=2.31.0
google-api-python-client>=2.100.0
google-genai>=1.0.0
apscheduler>=3.10.0
numpy>=1.24.0
matplotlib>=3.7.0
yt-dlp>=2025.1.1
beautifulsoup4>=4.12.0
playwright>=1.40.0
pycookiecheat>=0.6.0
